tidy-score: rename missnamed -> misnamed
